Duan Uys Design and basic logic of avrdude GUI
Updated list of compatiable boards. Gui reads from a file
and populates a combobox of supplied boards, with the uC information.

Once a board is selected, you can either flash or detect board with the two
simple buttons. Displays the f/w version of the intel hex file

Resolves:
See also: